Started to get a couple of tests, so moved those around so the classes don't clutter the unittest directory
