+++ tests/models/UnboundClass.php
@@ -1,0 +1,12 @@
+<?php
+namespace subdee\soapserver\tests\models;
+
+
+class UnboundClass
+{
+    /**
+     * @var string
+     * @soap
+     */
+    public $testVar;
+}
