+++ tests/models/UnboundedTestModel.php
@@ -1,0 +1,12 @@
+<?php
+namespace subdee\soapserver\tests\models;
+
+
+class UnboundedTestModel
+{
+    /**
+     * * @var UnboundClass[] {nillable = 0, minOccurs=1, maxOccurs = unbounded}
+     */
+    public $test = [];
+
+}
